sql_project/requirements.txt
alembic==1.16.1
arrow==1.3.0
bcrypt==4.3.0
binaryornot==0.4.4
blinker==1.9.0
cachelib==0.13.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Deprecated==1.2.18
dnspython==2.7.0
email_validator==2.2.0
environs==14.2.0
Flask==3.1.1
Flask-Bcrypt==1.0.1
Flask-Caching==2.3.1
Flask-DebugToolbar==0.16.0
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-MySQL==1.6.0
Flask-SQLAlchemy==3.1.1
Flask-Static-Digest==0.4.1
Flask-WTF==1.2.2
gevent==25.5.1
greenlet==3.2.3
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
limits==5.3.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==4.0.0
mdurl==0.1.2
mysql-connector-python==9.3.0
numpy==2.3.0
ordered-set==4.1.0
packaging==25.0
psycopg2-binary==2.9.10
pycparser==2.22
Pygments==2.19.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.4
rich==13.9.4
scikit-learn==1.7.0
scipy==1.15.3
setuptools==80.9.0
six==1.17.0
SQLAlchemy==2.0.41
supervisor==4.2.5
text-unidecode==1.3
threadpoolctl==3.6.0
types-python-dateutil==2.9.0.20250516
typing_extensions==4.14.0
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.17.2
WTForms==3.2.1
xgboost==3.0.2
zope.event==5.0
zope.interface==7.2
